refactorings in exercise 15 for cross-platform compatibility

+++ include/lib/mod01/m01_15.hpp
@@ -16,7 +16,7 @@ namespace concurr::m01_15 {
 
 template<typename Iter, typename T>
 void accumulate_block(Iter start, Iter end, T& result) {
-  result = std::accumulate(start, end, 0ULL);
+  result = std::accumulate(start, end, 0);
 }
 
 template<typename Iter, typename T>
@@ -27,18 +27,18 @@ T accumulate_list(Iter start, Iter end, T init) {
   
   auto constexpr min_elements_for_thread = 1'000'000;
   auto const max_threads = 
-    static_cast<uint64_t>(elements_total / min_elements_for_thread);
+    static_cast<int>(elements_total / min_elements_for_thread);
 //  std::cout << "Max number of threads to avoid heavy overload: " 
 //            << max_threads << "\n";
 
-  auto const allowed_threads = static_cast<uint64_t>(
+  auto const allowed_threads = static_cast<int>(
                                 std::thread::hardware_concurrency() > 0 ?
                                 std::thread::hardware_concurrency()     :
                                 2);
 //  std::cout << "Max number of threads to avoid oversubscription: "
 //            << allowed_threads << "\n";
 
-  auto const running_threads = std::min(allowed_threads, max_threads);
+  auto const running_threads = static_cast<int>(std::min(allowed_threads, max_threads));
 //  std::cout << "Runing threads: " << running_threads << "\n";
 
   auto results = std::vector<T>(running_threads);
@@ -47,7 +47,7 @@ T accumulate_list(Iter start, Iter end, T init) {
   auto const block_size = (elements_total + 1) / running_threads;
   auto block_start = Iter {start};
   
-  for (uint64_t i = 0; i < (running_threads - 1); ++i) {
+  for (int i = 0; i < (running_threads - 1); ++i) {
     auto block_end = Iter { block_start };
     std::advance(block_end, block_size);
     threads[i] = std::thread(
@@ -63,18 +63,18 @@ T accumulate_list(Iter start, Iter end, T init) {
 }
 
 void run() {
-  auto v = std::vector<uint64_t>(400'000'000);
+  auto v = std::vector<int>(400'000'000);
   std::iota(std::begin(v), std::end(v), 0);
 
   auto const time_start_par = detail::time_now();
-  auto const par_sum = accumulate_list(std::begin(v), std::end(v), 0ULL);
+  auto const par_sum = accumulate_list(std::begin(v), std::end(v), 0);
   auto const time_end_par = detail::time_elapsed(time_start_par);
   std::cout << "Parallel sum of list: " << par_sum << ". Elapsed: "
             << std::to_string(time_end_par) << "\n";
 
 
   auto const time_start_seq = detail::time_now();
-  auto const seq_sum = std::accumulate(std::begin(v), std::end(v), 0ULL);
+  auto const seq_sum = std::accumulate(std::begin(v), std::end(v), 0);
   auto const time_end_seq = detail::time_elapsed(time_start_seq);
   std::cout << "Sequential sum of list: " << seq_sum << ". Elapsed: "
             << std::to_string(time_end_seq) << "\n";
